requirements.txt
Pillow
slackclient